✨ Provide more configurations for page headers.
