+++ lib/calendar_view.dart
@@ -8,7 +8,6 @@ export './src/calendar_constants.dart';
 export './src/calendar_controller_provider.dart';
 export './src/calendar_event_data.dart';
 export './src/components/components.dart';
-export './src/components/safe_area_wrapper.dart';
 export './src/day_view/day_view.dart';
 export './src/enumerations.dart';
 export './src/event_arrangers/event_arrangers.dart';
