+++ lib/src/components/headers/day_page_header.dart
@@ -1,0 +1,44 @@
+// Copyright (c) 2021 Simform Solutions. All rights reserved.
+// Use of this source code is governed by a MIT-style license
+// that can be found in the LICENSE file.
+
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+
+import '../../constants.dart';
+import '../../style/header_style.dart';
+import '../../typedefs.dart';
+import 'calendar_page_header.dart';
+
+/// A header widget to display on day view.
+class DayPageHeader extends CalendarPageHeader {
+  /// A header widget to display on day view.
+  const DayPageHeader({
+    Key? key,
+    VoidCallback? onNextDay,
+    AsyncCallback? onTitleTapped,
+    VoidCallback? onPreviousDay,
+    StringProvider? dateStringBuilder,
+    required DateTime date,
+    @Deprecated("Use HeaderStyle to provide icon color")
+    Color iconColor = Constants.black,
+    @Deprecated("Use HeaderStyle to provide background")
+    Color backgroundColor = Constants.headerBackground,
+    HeaderStyle headerStyle = const HeaderStyle(),
+  }) : super(
+          key: key,
+          date: date,
+          // ignore_for_file: deprecated_member_use_from_same_package
+          backgroundColor: backgroundColor,
+          iconColor: iconColor,
+          onNextDay: onNextDay,
+          onPreviousDay: onPreviousDay,
+          onTitleTapped: onTitleTapped,
+          dateStringBuilder:
+              dateStringBuilder ?? DayPageHeader._dayStringBuilder,
+          headerStyle: headerStyle,
+        );
+
+  static String _dayStringBuilder(DateTime date, {DateTime? secondaryDate}) =>
+      "${date.day} - ${date.month} - ${date.year}";
+}
