+++ lib/src/components/month_view_components.dart
@@ -2,15 +2,12 @@
 // Use of this source code is governed by a MIT-style license
 // that can be found in the LICENSE file.
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import '../calendar_event_data.dart';
 import '../constants.dart';
 import '../extensions.dart';
-import '../style/header_style.dart';
 import '../typedefs.dart';
-import 'common_components.dart';
 
 class CircularCell extends StatelessWidget {
   /// Date of cell.
@@ -208,36 +205,6 @@ class FilledCell<T extends Object?> extends StatelessWidget {
       ),
     );
   }
-}
-
-class MonthPageHeader extends CalendarPageHeader {
-  /// A header widget to display on month view.
-  const MonthPageHeader({
-    Key? key,
-    VoidCallback? onNextMonth,
-    AsyncCallback? onTitleTapped,
-    VoidCallback? onPreviousMonth,
-    Color iconColor = Constants.black,
-    Color backgroundColor = Constants.headerBackground,
-    StringProvider? dateStringBuilder,
-    required DateTime date,
-    HeaderStyle headerStyle = const HeaderStyle(),
-  }) : super(
-          key: key,
-          date: date,
-          onNextDay: onNextMonth,
-          onPreviousDay: onPreviousMonth,
-          onTitleTapped: onTitleTapped,
-          // ignore_for_file: deprecated_member_use_from_same_package
-          backgroundColor: backgroundColor,
-          iconColor: iconColor,
-          dateStringBuilder:
-              dateStringBuilder ?? MonthPageHeader._monthStringBuilder,
-          headerStyle: headerStyle,
-        );
-
-  static String _monthStringBuilder(DateTime date, {DateTime? secondaryDate}) =>
-      "${date.month} - ${date.year}";
 }
 
 class WeekDayTile extends StatelessWidget {
